# cmd/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Bastio API
  description: AI Security Gateway API
  version: "0.1.0"
  license:
    name: FSL-1.1-ALv2
    url: https://fsl.software/FSL-1.1-ALv2.template.md
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /v1/overlays:
    get:
      operationId: listOverlays
      summary: List custom policies for this customer
      tags: [CustomPolicies]
      responses:
        "200":
          description: Overlays
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayList"
    post:
      operationId: createOverlay
      summary: Create a custom policy and its first draft version
      tags: [CustomPolicies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOverlayRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayCreateResult"
  /v1/overlays/from-template:
    post:
      operationId: createOverlayFromTemplate
      summary: Create a custom policy seeded from a built-in template
      tags: [CustomPolicies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayFromTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayCreateResult"
        "404":
          description: Template not found
  /v1/overlays/{id}:
    get:
      operationId: getOverlay
      summary: Fetch a custom policy with its active version and any loosening warnings
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteOverlay
      summary: Delete a custom policy and all of its versions
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /v1/overlays/{id}/versions:
    get:
      operationId: listOverlayVersions
      summary: List every version of a custom policy
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Versions, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayVersionList"
    post:
      operationId: createOverlayVersion
      summary: Append a new draft version
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOverlayVersionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayVersionResult"
  /v1/overlays/{id}/versions/{n}:
    get:
      operationId: getOverlayVersion
      summary: Fetch a specific version with loosening warnings
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: n
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayVersionDetail"
        "404":
          description: Not found
  /v1/overlays/{id}/versions/{n}/shadow:
    post:
      operationId: promoteOverlayVersionToShadow
      summary: Promote a draft version to shadow state
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: n
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayStateTransitionRequest"
      responses:
        "204":
          description: Promoted
        "404":
          description: Not found
        "409":
          description: Invalid state transition
  /v1/overlays/{id}/versions/{n}/activate:
    post:
      operationId: activateOverlayVersion
      summary: Activate a version, superseding the current active one
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: n
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayStateTransitionRequest"
      responses:
        "204":
          description: Activated
        "404":
          description: Not found
        "409":
          description: Invalid state transition
  /v1/overlays/{id}/versions/{n}/preview:
    post:
      operationId: previewOverlayVersion
      summary: Replay samples through the candidate effective profile and the current active, returning a diff
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: n
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayPreviewRequest"
      responses:
        "200":
          description: Preview result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayPreviewResult"
        "501":
          description: Preview runner not configured
  /v1/overlays/{id}/rollback:
    post:
      operationId: rollbackOverlay
      summary: Activate the most recently superseded version
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayStateTransitionRequest"
      responses:
        "204":
          description: Rolled back
        "409":
          description: No superseded version to roll back to
  /v1/overlays/{id}/audit:
    get:
      operationId: listOverlayAudit
      summary: Audit trail for a custom policy
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayAuditList"
  /v1/overlays/{id}/shadow-events:
    get:
      operationId: listOverlayShadowEvents
      summary: Shadow divergences recorded for this policy
      tags: [CustomPolicies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Shadow events, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayShadowEventList"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer sk-bastio-..."
  schemas:
...
    OverlayAccessRule:
      type: object
      required: [rule_type, value]
      description: Stored but not yet enforced at runtime.
      properties:
        rule_type:
          type: string
          enum: [ip_allowlist, ip_blocklist, geo_allow, geo_block]
        value:
          type: string
...
    OverlaySnapshot:
      type: object
      required: [schema_version]
      description: Delta-only snapshot — additions + overrides, never a copy of the core profile.
      properties:
        schema_version:
          type: integer
        additional_patterns:
          type: array
          items:
            $ref: "#/components/schemas/OverlayPatternRule"
        additional_access_rules:
          type: array
          items:
            $ref: "#/components/schemas/OverlayAccessRule"
        detector_overrides:
          $ref: "#/components/schemas/OverlayDetectorOverrides"
        plugin_detectors:
          type: array
          items:
            $ref: "#/components/schemas/OverlayPluginDetectorRef"
    Overlay:
      type: object
      required: [id, customer_id, name, description, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        proxy_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        active_version_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OverlayVersion:
      type: object
      required: [id, overlay_id, customer_id, version, state, snapshot, source, created_at]
      properties:
        id:
          type: string
          format: uuid
        overlay_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        version:
          type: integer
        state:
          type: string
          enum: [draft, shadow, active, superseded]
        snapshot:
          $ref: "#/components/schemas/OverlaySnapshot"
        source:
          type: string
          description: "manual | template:<slug> | import"
        commit_message:
          type: string
        created_by:
          type: string
        shadow_started_at:
          type: string
          format: date-time
          nullable: true
        activated_at:
          type: string
          format: date-time
          nullable: true
        superseded_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    OverlayWarning:
      type: object
      required: [detector, field, from, to, message]
      description: Flags a snapshot override that weakens security relative to the shipped defaults.
      properties:
        detector:
          type: string
        field:
          type: string
          enum: [threshold, strategy, action]
        from:
          type: string
        to:
          type: string
        message:
          type: string
    OverlayPreviewSample:
      type: object
      required: [content]
      properties:
        content:
          type: string
    OverlayPreviewSampleResult:
      type: object
      required: [content, active_action, candidate_action, active_block, candidate_block]
      properties:
        content:
          type: string
        active_action:
          type: string
        candidate_action:
          type: string
        active_block:
          type: boolean
        candidate_block:
          type: boolean
        divergence:
          type: string
    OverlayPreviewSummary:
      type: object
      required: [total, matching, would_block, would_allow, threshold_diff]
      properties:
        total:
          type: integer
        matching:
          type: integer
        would_block:
          type: integer
        would_allow:
          type: integer
        threshold_diff:
          type: integer
    OverlayPreviewResult:
      type: object
      required: [results, summary]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/OverlayPreviewSampleResult"
        summary:
          $ref: "#/components/schemas/OverlayPreviewSummary"
    CreateOverlayRequest:
      type: object
      required: [snapshot]
      properties:
        name:
          type: string
        description:
          type: string
        proxy_id:
          type: string
          format: uuid
        snapshot:
          $ref: "#/components/schemas/OverlaySnapshot"
        commit_message:
          type: string
        source:
          type: string
    CreateOverlayVersionRequest:
      type: object
      required: [snapshot]
      properties:
        snapshot:
          $ref: "#/components/schemas/OverlaySnapshot"
        commit_message:
          type: string
        source:
          type: string
    OverlayStateTransitionRequest:
      type: object
      properties:
        reason:
          type: string
    OverlayFromTemplateRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        proxy_id:
          type: string
          format: uuid
        commit_message:
          type: string
    OverlayPreviewRequest:
      type: object
      required: [samples]
      properties:
        samples:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/OverlayPreviewSample"
        proxy_id:
          type: string
          format: uuid
    OverlayList:
      type: object
      required: [overlays]
      properties:
        overlays:
          type: array
          items:
            $ref: "#/components/schemas/Overlay"
    OverlayCreateResult:
      type: object
      required: [overlay, version]
      properties:
        overlay:
          $ref: "#/components/schemas/Overlay"
        version:
          $ref: "#/components/schemas/OverlayVersion"
        template:
          type: string
          description: Present when created via /overlays/from-template.
    OverlayDetail:
      type: object
      required: [overlay]
      properties:
        overlay:
          $ref: "#/components/schemas/Overlay"
        active_version:
          $ref: "#/components/schemas/OverlayVersion"
        active_warnings:
          type: array
          items:
            $ref: "#/components/schemas/OverlayWarning"
    OverlayVersionList:
      type: object
      required: [versions]
      properties:
        versions:
          type: array
          items:
            $ref: "#/components/schemas/OverlayVersion"
    OverlayVersionResult:
      type: object
      required: [version]
      properties:
        version:
          $ref: "#/components/schemas/OverlayVersion"
    OverlayVersionDetail:
      type: object
      required: [version]
      properties:
        version:
          $ref: "#/components/schemas/OverlayVersion"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/OverlayWarning"
    OverlayAuditList:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/OverlayAuditEntry"
    OverlayShadowEventList:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/OverlayShadowEvent"
...